Prevents infinite recursion while visiting by memoizing a placeholder (as @JoeOsborn noted, sort of a future) on visit, and filling in its value later on.
This is still incorrect, as recursion causes an incorrect value to be uses as the value of the outermost recursive node. Some thought needs to be given to how to serialize given that, perhaps with lazy results or with (further) placeholders.

I’m fairly confident that the solution should make it possible to implement some fairly diverse structured algorithms over the (arbitrarily cyclic) graph, though.