Adds a parameter to each of the visitor methods to pass self in.
Object algebras can capture the identity of the visited objects, but only by maintaining a stack of visited nodes in parallel to the call stack. This makes it unnecessarily difficult to express several algorithms involving parser combinators, but also allows you to use the algebra protocol to implement a factory.

Since I have no particular need of factories or a solution to the expression problem, and I do have a vested interest in convenient and efficient recursive operations on parser combinators, I opted to make the visitor protocol a little more Visitor-like.

It still retains several interesting benefits of the lazy algebra approach I had taken before, however:

- encapsulation is preserved
- specialization of operations by type does not need to resort to -isKindOfClass:
- lazy recursion is still direct