Delayed sets hash to 0 before they’re forced.
This is consistent with -[NSMutableSet hash] if you consider the delay
set as being mutable (over time). It would be nice to parameterize it
by the actual count, but given that this is being populated by
arbitrary reductions (which could trivially incorporate flatten
maps/filters), I’m considering it to be out of scope for the moment.